Service rewrite (#121)
* Refactoring

* Refactoring

* Refactoring

* Refactoring

* Rewrite

* Move files to Stations directory

* Update application.properties

* Rewrite

* refactor

* Change to Long

* Refactoring

* Refactoring

* Update Dockerfile.jvm

* Refactoring

* Refactoring

* Refactoring

* Refactoring

* Refactoring

* Rewrite

* Move files to Stations directory

* Update application.properties

* Rewrite

* refactor

* Change to Long

* Refactoring

* Refactoring

* Update Dockerfile.jvm

* Refactoring

* Remove json files

* Remove .idea

+++ src/main/java/si/deisinger/providers/model/implera/ImpleraLocations.java
@@ -6,21 +6,21 @@ import java.util.List;
 
 public class ImpleraLocations {
 
-	@JacksonXmlElementWrapper(useWrapping = false)
-	public List<marker> marker;
+    @JacksonXmlElementWrapper(useWrapping = false)
+    public List<marker> marker;
 
-	public static class marker {
-		public int id;
-		public String name;
-		public String address;
-		public String town;
-		public double lat;
-		public double lng;
-		public int deluje;
-		public Object cena_KW;
-		public Object reklamni_tekst_baner;
-		public Object cena_na_vklop;
-		public Object opis;
-		public String opis_tip_prikljucka_vticnice;
-	}
+    public static class marker {
+        public Long id;
+        public String name;
+        public String address;
+        public String town;
+        public double lat;
+        public double lng;
+        public int deluje;
+        public Object cena_KW;
+        public Object reklamni_tekst_baner;
+        public Object cena_na_vklop;
+        public Object opis;
+        public String opis_tip_prikljucka_vticnice;
+    }
 }
